Statistical Image Segmentation on GPU
The algorithm uses the advantages of GPUs to accelerate RGB image segmentation and It has been implemented in CUDA C++
environment. Statistical criteria for image data and use of GPUs are considered. The algorithm is the efficiently segmented
color image with better acceleration. The parallel algorithm is carried-out on GPU with C++ program language and used to segmenting the RGB color images.